requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
click==8.2.0
ecdsa==0.19.0
fastapi==0.110.0
h11==0.14.0
idna==3.7
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.10.2
pydantic-core==2.18.1
python-jose==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
uvicorn==0.24.0
sqlalchemy==2.0.23